Refactoring of Storage - Tags
