+++ src/main/java/com/iota/iri/service/storage/Storage.java
@@ -21,7 +21,6 @@ public class Storage extends AbstractStorage {
 	
 	private static final Logger log = LoggerFactory.getLogger(Storage.class);
 
-    private static final String TAGS_FILE_NAME = "tags.iri";
     private static final String APPROVERS_FILE_NAME = "approvers.iri";
     private static final String SCRATCHPAD_FILE_NAME = "scratchpad.iri";
 
@@ -29,10 +28,6 @@ public class Storage extends AbstractStorage {
     
     public static int numberOfApprovedTransactionsToStore;
     
-    private FileChannel tagsChannel;
-    private final ByteBuffer[] tagsChunks = new ByteBuffer[MAX_NUMBER_OF_CHUNKS];
-    private volatile long tagsNextPointer = SUPER_GROUPS_SIZE;
-
     private FileChannel approversChannel;
     private final ByteBuffer[] approversChunks = new ByteBuffer[MAX_NUMBER_OF_CHUNKS];
     private volatile long approversNextPointer = SUPER_GROUPS_SIZE;
@@ -51,6 +46,7 @@ public class Storage extends AbstractStorage {
     private StorageTransactions storageTransactionInstance = StorageTransactions.instance();
     private StorageBundle storageBundleInstance = StorageBundle.instance();
     private StorageAddresses storageAddressesInstance = StorageAddresses.instance();
+    private StorageTags storageTags = StorageTags.instance();
     
     @Override
     public synchronized void init() throws IOException {
@@ -58,37 +54,8 @@ public class Storage extends AbstractStorage {
     	storageTransactionInstance.init();
     	storageBundleInstance.init();
     	storageAddressesInstance.init();
-    	
-        tagsChannel = FileChannel.open(Paths.get(TAGS_FILE_NAME), StandardOpenOption.CREATE, StandardOpenOption.READ, StandardOpenOption.WRITE);
-        tagsChunks[0] = tagsChannel.map(FileChannel.MapMode.READ_WRITE, 0, SUPER_GROUPS_SIZE);
-        final long tagsChannelSize = tagsChannel.size();
-        while (true) {
-
-            if ((tagsNextPointer & (CHUNK_SIZE - 1)) == 0) {
-                tagsChunks[(int)(tagsNextPointer >> 27)] = tagsChannel.map(FileChannel.MapMode.READ_WRITE, tagsNextPointer, CHUNK_SIZE);
-            }
-
-            if (tagsChannelSize - tagsNextPointer > CHUNK_SIZE) {
-                tagsNextPointer += CHUNK_SIZE;
-
-            } else {
-
-                tagsChunks[(int) (tagsNextPointer >> 27)].get(mainBuffer);
-                boolean empty = true;
-                for (final int value : mainBuffer) {
-                    if (value != 0) {
-                        empty = false;
-                        break;
-                    }
-                }
-                if (empty) {
-                    break;
-                }
-
-                tagsNextPointer += CELL_SIZE;
-            }
-        }
-
+    	storageTags.init();
+        
         approversChannel = FileChannel.open(Paths.get(APPROVERS_FILE_NAME), StandardOpenOption.CREATE, StandardOpenOption.READ, StandardOpenOption.WRITE);
         approversChunks[0] = approversChannel.map(FileChannel.MapMode.READ_WRITE, 0, SUPER_GROUPS_SIZE);
         final long approversChannelSize = approversChannel.size();
@@ -137,12 +104,8 @@ public class Storage extends AbstractStorage {
         	storageTransactionInstance.shutdown();
         	storageBundleInstance.shutdown();
         	storageAddressesInstance.shutdown();
-
-            for (int i = 0; i < MAX_NUMBER_OF_CHUNKS && tagsChunks[i] != null; i++) {
-            	log.info("Flushing tags chunk #" + i);
-                flush(tagsChunks[i]);
-            }
-
+        	storageTags.shutdown();
+        	
             for (int i = 0; i < MAX_NUMBER_OF_CHUNKS && approversChunks[i] != null; i++) {
             	log.info("Flushing approvers chunk #" + i);
                 flush(approversChunks[i]);
@@ -151,9 +114,7 @@ public class Storage extends AbstractStorage {
             log.info("DB successfully flushed");
 
             try {
-                tagsChannel.close();
                 approversChannel.close();
-
             } catch (final Exception e) {
             	log.error("Catched Exception whilst shutting down:", e);
             }
@@ -248,79 +209,6 @@ public class Storage extends AbstractStorage {
         analyzedTransactionsFlags.put(analyzedTransactionsFlagsCopy);
     }
 
-    public synchronized long tagPointer(final byte[] hash) {
-
-        long pointer = ((hash[0] + 128) + ((hash[1] + 128) << 8)) << 11;
-        for (int depth = 2; depth < Transaction.TAG_SIZE; depth++) {
-
-            ((ByteBuffer) tagsChunks[(int)(pointer >> 27)].position((int)(pointer & (CHUNK_SIZE - 1)))).get(mainBuffer);
-
-            if (mainBuffer[Transaction.TYPE_OFFSET] == GROUP) {
-
-                if ((pointer = value(mainBuffer, (hash[depth] + 128) << 3)) == 0) {
-                    return 0;
-                }
-
-            } else {
-
-                for (; depth < Transaction.TAG_SIZE; depth++) {
-                    if (mainBuffer[Transaction.HASH_OFFSET + depth] != hash[depth]) {
-                        return 0;
-                    }
-                }
-
-                return pointer;
-            }
-        }
-
-        throw new IllegalStateException("Corrupted storage");
-    }
-
-    public synchronized List<Long> tagTransactions(final long pointer) {
-
-        final List<Long> tagTransactions = new LinkedList<>();
-
-        if (pointer != 0) {
-
-            ((ByteBuffer) tagsChunks[(int) (pointer >> 27)].position((int) (pointer & (CHUNK_SIZE - 1)))).get(mainBuffer);
-            int offset = ZEROTH_POINTER_OFFSET - Long.BYTES;
-            while (true) {
-
-                while ((offset += Long.BYTES) < CELL_SIZE - Long.BYTES) {
-
-                    final long transactionPointer = value(mainBuffer, offset);
-                    if (transactionPointer == 0) {
-
-                        break;
-
-                    } else {
-
-                        tagTransactions.add(transactionPointer);
-                    }
-                }
-                if (offset == CELL_SIZE - Long.BYTES) {
-
-                    final long nextCellPointer = value(mainBuffer, offset);
-                    if (nextCellPointer == 0) {
-
-                        break;
-
-                    } else {
-
-                        ((ByteBuffer) tagsChunks[(int) (nextCellPointer >> 27)].position((int) (nextCellPointer & (CHUNK_SIZE - 1)))).get(mainBuffer);
-                        offset = -Long.BYTES;
-                    }
-
-                } else {
-
-                    break;
-                }
-            }
-        }
-
-        return tagTransactions;
-    }
-
     public synchronized long approveePointer(final byte[] hash) {
 
         long pointer = ((hash[0] + 128) + ((hash[1] + 128) << 8)) << 11;
@@ -376,7 +264,6 @@ public class Storage extends AbstractStorage {
                         ((ByteBuffer) approversChunks[(int) (nextCellPointer >> 27)].position((int) (nextCellPointer & (CHUNK_SIZE - 1)))).get(mainBuffer);
                         offset = -Long.BYTES;
                     }
-
                 } else {
                     break;
                 }
@@ -384,19 +271,6 @@ public class Storage extends AbstractStorage {
         }
 
         return approveeTransactions;
-    }
-
-    private void appendToTags() {
-
-        ((ByteBuffer) tagsChunks[(int)(tagsNextPointer >> 27)].position((int)(tagsNextPointer & (CHUNK_SIZE - 1)))).put(mainBuffer);
-        if (((tagsNextPointer += CELL_SIZE) & (CHUNK_SIZE - 1)) == 0) {
-
-            try {
-                tagsChunks[(int)(tagsNextPointer >> 27)] = tagsChannel.map(FileChannel.MapMode.READ_WRITE, tagsNextPointer, CHUNK_SIZE);
-            } catch (final IOException e) {
-            	log.error("Caught exception on appendToTags:", e);
-            }
-        }
     }
 
     private void appendToApprovers() {
@@ -421,120 +295,8 @@ public class Storage extends AbstractStorage {
         numberOfApprovedTransactionsToStore = 0;
 
         StorageBundle.instance().updateBundle(transactionPointer, transaction);
-
         StorageAddresses.instance().updateAddresses(transactionPointer, transaction);
-
-        for (int i = 0; i < Transaction.TAG_SIZE; i++) {
-
-            if (transaction.tag[i] != 0) {
-
-                long pointer = ((transaction.tag[0] + 128) + ((transaction.tag[1] + 128) << 8)) << 11, prevPointer = 0;
-                for (int depth = 2; depth < Transaction.TAG_SIZE; depth++) {
-
-                    ((ByteBuffer) tagsChunks[(int)(pointer >> 27)].position((int)(pointer & (CHUNK_SIZE - 1)))).get(mainBuffer);
-
-                    if (mainBuffer[Transaction.TYPE_OFFSET] == GROUP) {
-
-                        prevPointer = pointer;
-                        if ((pointer = value(mainBuffer, (transaction.tag[depth] + 128) << 3)) == 0) {
-
-                            setValue(mainBuffer, (transaction.tag[depth] + 128) << 3, tagsNextPointer);
-                            ((ByteBuffer) tagsChunks[(int)(prevPointer >> 27)].position((int)(prevPointer & (CHUNK_SIZE - 1)))).put(mainBuffer);
-
-                            System.arraycopy(ZEROED_BUFFER, 0, mainBuffer, 0, CELL_SIZE);
-                            mainBuffer[Transaction.TYPE_OFFSET] = FILLED_SLOT;
-                            System.arraycopy(transaction.tag, 0, mainBuffer, 8, Transaction.TAG_SIZE);
-                            setValue(mainBuffer, ZEROTH_POINTER_OFFSET, transactionPointer);
-                            appendToTags();
-
-                            break;
-                        }
-
-                    } else {
-
-                        boolean sameTag = true;
-
-                        for (int j = depth; j < Transaction.TAG_SIZE; j++) {
-
-                            if (mainBuffer[Transaction.HASH_OFFSET + j] != transaction.tag[j]) {
-
-                                final int differentHashByte = mainBuffer[Transaction.HASH_OFFSET + j];
-
-                                ((ByteBuffer) tagsChunks[(int)(prevPointer >> 27)].position((int)(prevPointer & (CHUNK_SIZE - 1)))).get(mainBuffer);
-                                setValue(mainBuffer, (transaction.tag[depth - 1] + 128) << 3, tagsNextPointer);
-                                ((ByteBuffer) tagsChunks[(int)(prevPointer >> 27)].position((int)(prevPointer & (CHUNK_SIZE - 1)))).put(mainBuffer);
-
-                                for (int k = depth; k < j; k++) {
-
-                                    System.arraycopy(ZEROED_BUFFER, 0, mainBuffer, 0, CELL_SIZE);
-                                    setValue(mainBuffer, (transaction.tag[k] + 128) << 3, tagsNextPointer + CELL_SIZE);
-                                    appendToTags();
-                                }
-
-                                System.arraycopy(ZEROED_BUFFER, 0, mainBuffer, 0, CELL_SIZE);
-                                setValue(mainBuffer, (differentHashByte + 128) << 3, pointer);
-                                setValue(mainBuffer, (transaction.tag[j] + 128) << 3, tagsNextPointer + CELL_SIZE);
-                                appendToTags();
-
-                                System.arraycopy(ZEROED_BUFFER, 0, mainBuffer, 0, CELL_SIZE);
-                                mainBuffer[Transaction.TYPE_OFFSET] = FILLED_SLOT;
-                                System.arraycopy(transaction.tag, 0, mainBuffer, 8, Transaction.TAG_SIZE);
-                                setValue(mainBuffer, ZEROTH_POINTER_OFFSET, transactionPointer);
-                                appendToTags();
-
-                                sameTag = false;
-
-                                break;
-                            }
-                        }
-
-                        if (sameTag) {
-
-                            int offset = ZEROTH_POINTER_OFFSET;
-                            while (true) {
-
-                                while ((offset += Long.BYTES) < CELL_SIZE - Long.BYTES && value(mainBuffer, offset) != 0) {
-
-                                    // Do nothing
-                                }
-                                if (offset == CELL_SIZE - Long.BYTES) {
-
-                                    final long nextCellPointer = value(mainBuffer, offset);
-                                    if (nextCellPointer == 0) {
-
-                                        setValue(mainBuffer, offset, tagsNextPointer);
-                                        ((ByteBuffer) tagsChunks[(int)(pointer >> 27)].position((int)(pointer & (CHUNK_SIZE - 1)))).put(mainBuffer);
-
-                                        System.arraycopy(ZEROED_BUFFER, 0, mainBuffer, 0, CELL_SIZE);
-                                        setValue(mainBuffer, 0, transactionPointer);
-                                        appendToTags();
-
-                                        break;
-
-                                    } else {
-
-                                        pointer = nextCellPointer;
-                                        ((ByteBuffer) tagsChunks[(int)(pointer >> 27)].position((int)(pointer & (CHUNK_SIZE - 1)))).get(mainBuffer);
-                                        offset = -Long.BYTES;
-                                    }
-
-                                } else {
-
-                                    setValue(mainBuffer, offset, transactionPointer);
-                                    ((ByteBuffer) tagsChunks[(int)(pointer >> 27)].position((int)(pointer & (CHUNK_SIZE - 1)))).put(mainBuffer);
-
-                                    break;
-                                }
-                            }
-                        }
-
-                        break;
-                    }
-                }
-
-                break;
-            }
-        }
+        StorageTags.instance().updateTags(transactionPointer, transaction);
 
         updateApprover(transaction.trunkTransaction, transactionPointer);
         if (transaction.branchTransactionPointer != transaction.trunkTransactionPointer) {
@@ -640,19 +402,7 @@ public class Storage extends AbstractStorage {
         }
     }
     
-    // methods helpes
-    
-    public byte[] mainBuffer() {
-		return mainBuffer;
-	}
-    
-    public byte[] auxBuffer() {
-		return auxBuffer;
-	}
-    
-    public byte[] zeroedBuffer() {
-		return ZEROED_BUFFER;
-	}
+    // methods helper
     
     private static Storage instance = new Storage();
